Update patch set 12
Patch Set 12: Code-Review+2

I don't have any strong opinion about the local path. There are 2 contradicting arguments here - preserve the path that user provides and have ability to run on multiple devices. I'll leave it as it is now. If this proves to be a problem, we can change it later.

Patch-set: 12
Label: Code-Review=+2