Update patch set 3
Patch Set 3:

(3 comments)

Also from this script, seems the implementation depends on host side adb shell command, could we change to make the test to run on device side and output the result, and host side only to parse the output and upload the result to lava(or other framework)?

If we could make the test script run device side, then the test could be run at anytime as soon as the device is up. Otherwise like now, to run the test we depends on this repository. And if there is a V3 in the future, we might need to change again.

Patch-set: 3