Update patch set 3
Patch Set 3:

(3 comments)

> (2 comments)
 > 
 > YongQin, I don't understand the comment about running the script on
 > device. The test (stringbench) runs on device. The host side only
 > parses the result. So it's exactly as you request, so what is the
 > problem?
what I like is to push this stringbench.sh to device, and run it on device side, and output the better format result, on the host side just doing the things could not be done on target side.

This way, the stringbench.sh don't need to change if there is LavaV3 in future.
Otherwise, like if there is a problem that adb connection does not work well on lava, and need to change to run tests via serial, then no need to change the stringbench.sh script again I think.

I know V2 depends on the adb connection, just in case.

Also just thought of one thing, how is it supposed to run same test for multiple target devices on one host? like run this stringbench test on hikey and x15 at the same time.

Patch-set: 3