Update patch set 3
Patch Set 3:

(2 comments)

> This way, the stringbench.sh don't need to change if there is LavaV3 in future. Otherwise, like if there is a problem that adb connection does not work well on lava, and need to change to run tests via serial, then no need to change the stringbench.sh script again I think.

I don't like the idea of relying on scripts that are included at the build time. In this case, the script is so easy I don't think there is no need to copy it to target before executing test.

> I know V2 depends on the adb connection, just in case.

> Also just thought of one thing, how is it supposed to run same test for multiple target devices on one host? like run this stringbench test on hikey and x15 at the same time.

Why do you think it's the case here? adb only connects to 1 device at a time. So what is the question about?

Patch-set: 3