Update patch set 3
Patch Set 3:

> , I will not insist on it. just the test scripts will depend more and more on host side tools.

The request from Vishal was to provide the same level of 'service' as we used to have with lava-android-test. It depended on the host side tools as well, so I don't think this is a problem. 

> On LAVA, it's true. but how about on local side? But we are considering to run the test on local side as well, right? To save time, I would want to run same test on different devices to compare the result.

Yes, this is exactly the idea. I guess you can't run on all devices 'at the same time' but you surely can run one after another and than compare the results. There are probably missing bits here (recording environment, result comparison tools). IMHO the way to run on different devices would be just to change ANDROID_SERIAL variable and re-run the test (from the script r from the test plan). I'll create a ticket on that.

Patch-set: 3
Label: Code-Review=0