Update patch set 2
Patch Set 2:

I'm not 100% sure, but I suspect there might be a problem with CTS in case more than 1 device is connected. The ANDROID_SERIAL is not exported in this test (no call to initialize_adb) and adb is called in subshells from python. Please check if cts works when more than 1 device is connected?

Patch-set: 2
Label: Code-Review=0