android-test-lib: add parse_common_args()
Moving the parsing for common args to android-test-lib can avoid code
duplication.

On the other hand, except the common args, if the test script support
additional args, IMO, we should put them together and do the paring in
the script. Actually, with the capability to handle illegal input, I
don't see an easy way to split the args. And make the code easy to read
also is a priority.

Change-Id: I6ea7ac4a624f6accb509aaf4e735cd8e190fcc36
Signed-off-by: Chase Qi <chase.qi@linaro.org>