Update patch set 1
Patch Set 1:

I think there are 2 cases here:
1. running deprecated stuff from test-definitions
At some point I'm planning to delete these tests. Could you give me the list of the ones that are not yet migrated to 'automated/' and you're using them?

2. running an external repository stuff
This gets tricky. So far the idea was to save some space. In case of 3rd party repo that isn't known, we would probably need to do the same thing as LAVA does - clone the whole thing. So space saving doesn't apply. If that's OK for the SSH use case, I'm OK with that.

Patch-set: 1
Label: Code-Review=0