test-runner: add support for parse pattern
LAVA V2 is able to catch and save result lines that send to stdout
instantly. For tests that run for a long time and generates a plenty of
results as it runs, like kselftest and ltp tests, LAVA's live result
collection feature could be useful.

In Linaro Test Suite, to support completely local run and reduce the
complexity of log parser debugging, we parse test log in test script
and save results in LAVA's format to result.txt file. Then use
'send-to-lava.sh [path]/result.txt' to send result lines to stdout after
test finished. This should work for most test cases, but maybe not good
enough for tests mentioned above as the tests may crash or hang or cause
Kernel panic or system halt during test run, in those cases, the result
sending line has no chance to get executed.

Using parse pattern in test definition file is one way to enable LAVA's
live result collection feature. This patch is used to add parse pattern
support in test-runner so that we can run the test cases using parse
pattern on the local. It also could be useful for pattern debugging on
the local.

Change-Id: I5756af84f653ab00fc52c62ff888b71fe8817944
Signed-off-by: Chase Qi <chase.qi@linaro.org>