Update patch set 3
Patch Set 3:

hmm. why would we need something like that? I would tend to think that the 'test plan' is the master and gets to decide what to run. test-runner isn't supposed to try to be smarter than whoever wrote the test plan. maybe we could print a warning for duplicate tests instead of blindly skipping them.

Patch-set: 3