Update patch set 3
Patch Set 3:

I asked the same question before. It should be pretty hard to mock the same pattern that is used in parse_stdout(). It's not impossible, but very unlikely IMHO.

kselftest might be used as example. I'll try to run some tests locally today and show the output.

Patch-set: 3