test-runner: Report environmental information
Add a dictionary key named 'environment' that identifies the host running the
tests, whether that host is local or remote. For example, on an ubuntu system,
the following data structure will be added to each result in results.json:

{
    "environment": {
        "kernel": "4.8.0-49-generic",
        "packages": [
            "0ad-0.0.20-3",
            "0ad-data-0.0.20-1",
            ...
            "zlib1g-1:1.2.8.dfsg-2ubuntu5",
            "zlib1g-dev-1:1.2.8.dfsg-2ubuntu5"
        ],
        "linux_distribution": "ubuntu",
        "uname": "Linux zygarde 4.8.0-49-generic #52-Ubuntu SMP Thu Apr 20 09:38:39 UTC 2017 x86_64 x86_64 x86_64 GNU/Linux"
    }
}

Support exists for Ubuntu, Debian, CentOS, and Fedora. Other distributions will
fail unless support is explicitly added. Environmental data is added
automatically in each test and is not optional.

This change led to a slight refactor in the way remote commands were being
handled. Instead of percolating remote vs local logic everywhere, run_command
will run an arbitrary shell command either locally or on the target host, and
return the output. This allows all shell commands, whether local or remote, to
be handled consistently.

Change-Id: Ia94695751a04a7fc862a299526d33784c57a9893
Signed-off-by: Dan Rue <dan.rue@linaro.org>