test-runner: Report environmental information
Add a dictionary key named 'environment' that identifies the host running the
tests, whether that host is local or remote. For example, on a HiSilicon D03
system, the following data structure will be added to each result in
results.json:

{
    "bios_version": "Hisilicon D03 UEFI 16.12 Release",
    "board_name": "D03",
    "board_vendor": "Huawei",
    "kernel": "4.9.0-20.gitedc2a1c.linaro.aarch64",
    "linux_distribution": "centos",
    "packages": [
        "GeoIP-1.5.0-11.el7",
        "NetworkManager-1.4.0-20.el7_3",
        ...
        "yum-plugin-fastestmirror-1.1.31-40.el7",
        "zlib-1.2.7-17.el7"
    ],
    "uname": "Linux localhost.localdomain 4.9.0-20.gitedc2a1c.linaro.aarch64 #1 SMP Wed Dec 14 17:50:15 UTC 2016 aarch64 aarch64 aarch64 GNU/Linux"
}

Support exists for Ubuntu, Debian, CentOS, and Fedora. Other distributions will
print a warning and report an empty package list.  Environmental data is added
automatically in each test and may be skipped by passing -e or
--skip-environment.

This change led to a slight refactor in the way remote commands were being
handled. Instead of percolating remote vs local logic everywhere, run_command
will run an arbitrary shell command either locally or on the target host, and
return the output. This allows all shell commands, whether local or remote, to
be handled consistently.

Change-Id: Ia94695751a04a7fc862a299526d33784c57a9893
Signed-off-by: Dan Rue <dan.rue@linaro.org>