Update patch set 4
Patch Set 4: Code-Review+2

(2 comments)

This looks like a good implementation to me.

The awkwardness in squad_metadata (anywhere you see a [-1]), is a result of results.json being too flat. I think we could improve this by changing the structure of results.yaml to be more representative of the actual data.

Patch-set: 4
Label: Code-Review=+2