Update patch set 2
Patch Set 2: -Code-Review

Fathi,
Thanks for the review.

 > wrt the change itself, I don't understand why we need to know where
 > libhugetlbfs is installed in the first place? 

Because This test script should work on OE and Debian based builds.

As you know on OE, Detect pre-built path and run test. When testing on EPR builds (Debian or CentOS) - it is NOT pre-installed so test downloads and builds libhugetlbfs test suite on target.

 > If it is about
 > detecting if it's pre-installed, there's much saner way to achieve
 > it. Note that the following can help:
 > * find /usr/*/libhugetlbfs -type f -name run_tests.py
 > * cd /usr/*/libhugetlbfs
 > 
 > In other words, wildcard can help you here.

Wildcard is helpful.
Thanks.

Patch-set: 2