linux: ltp: skip min_free_kbytes on RPK
By default, Debian installer will set the size of swap partition to the
size of memory capacity. min_free_kbytes needs to occupy both memory and
swap spaces. When memory size is large, this test takes hours to finish.
For example, on D05 that 128G memory installed, this test takes more
then 3 hours. It can be finished, but just too slow. It is possbile to
customize partitioning steps to reduce swap partition size for this
single test, but that is probably not wanted.

Change-Id: I99e397f5277e49c9e01f7d73f34f126f5a8e3764
Signed-off-by: Chase Qi <chase.qi@linaro.org>