Update patch set 2
Patch Set 2: Code-Review+2

That looks good. It might get a bit tricky if the tarball doesn't have 'kselftest' top level directory. I think we should test for directory existence and exit gracefully if it's not there. So far we were building the tarballs ourselves. From this patch it won't be the case any more.

Patch-set: 2
Label: Code-Review=+2