Update patch set 1
Patch Set 1:

> > Could tests under /automated/android be used on the lava
 > production instance now?
 > 
 > The path doesn't matter that much. It's the type of test that
 > matters.
Understand that, just the problem is that, if I move these 2 files under /automated/android without any change, then the test implementations there will be messed up I think. This implementations of bootstat could only be run with V1, and others could only be run with V2.

But if I update it as others in /automated/android, then it will not be able to run on V1 which is the production instance status we are using.

what I want is to run it for the android build, what's your suggestions for that:
1. no update on this change, so that it could be run on lava production at the moment
2. move to /automated/android,  but no update on this change, so that it could be run  on lava production at the moment.
3. move to /automated/android, update it as others under there. but it could not be run on lava now at the moment. (need to wait lava deploys the v2 for hikey and other platforms on production)

Patch-set: 1
Label: Code-Review=0