Update patch set 3
Patch Set 3:

> YongQin,
 > Please move it to automated/android. It doesn't prevent you from
 > running in v1. I think we can marry both versions here. The
 > bootstat.sh script will be the same in both versions (or will have
 > minor changes). The yaml file you proposed is v1 only. So let's
 > rename it to bootstat_v1.yaml. Later we can create bootstat_v2.yaml
 > (or simply bootstat.yaml) and add v2 support (running from LXC).
 > 
renamed the bootstat.sh script to device-script.sh as well like what is there for other test.

 > Please also add a comment in the yaml file stating that it's meant
 > for LAVA v1 only.
 > 
 > I think you're missing the pattern in the yaml.
I use the output_test_result common function which internally uses lava-test-case command for the test result recording. personally I don't like the pattern method in yaml file, it's not easy to read the regular expression, also not possible to handle multiple cases(or say need to be more complex on the regular expression).

I see the output_test_result function is already defined in boottime test, but I think it's more like a common function might be use by other tests as well, so I still source the old common file under android/script in the device-script.sh, this part needs to be updated when add the V2 support.

I don't like to copy the same function to multiple places, my suggestion is to create a common
script for the device side, and each device-script.sh could source it. but you could discuss how to do that in your team, and I will follow your decisions.

Patch-set: 3
Label: Code-Review=0