Update patch set 2
Patch Set 2:

(3 comments)

> Regarding the binaries in this patch:
 > - Why can't we compile them as a part of the test?

This was discussed before. The major reason is we want to support distributions that don't have toolchian installed and don't have package management support like the OE based builds Linaro building.

This test case is required by PRB test plan(shared with you), it is not EE only. If we can take it as ERP only, I am more then happy to do the compiling at runtime. I will take to Milosz about this at 1:1 sync.

 > - How do we know how they were compiled?
 > - How do we know what version of source was used?

If we decide to use the binaries, I will include the commit ID and compile commands in the README file.

 > - How will we update them in the future?

It isn't clear yet. IMO, for stress testing purpose, we may don't need the latest code, we need a known good stable base. 

 > - What about x86 platforms (are they in scope for test-definitions?)
x86 is not supported yet.

Patch-set: 2