android: tradefed: comment out wait_homescreen()
wait_homescreen() searches logcat output for
'Displayed com.android.launcher', but the log might be washed away when
a lot of logs generated after it. When the function not executed in
time, error occurs. This has been observer several times on lkft
testing. Refer to the following link:
    https://lkft.validation.linaro.org/scheduler/job/18918#L4721

We are already using wait_boot_completed() to check boot status, lets
comment out wait_homescreen() and see if wait_boot_completed() is
sufficient.

Change-Id: I930ac469cc3ea0d726f676c84413c91f6c83e700
Signed-off-by: Chase Qi <chase.qi@linaro.org>