android: fix adb root
'adb root' and then 'sleep 10' is a work around for some versions of
adb. In LAVA LXC, with that version adb, 'adb root' doesn't return, but
it actually restarted as root. Without 'sleep 10', 'adb wait-for-device'
and 'adb devices' may still look good, but the following adb commands
will show 'no device found'. Refer to the below job.

  https://lkft.validation.linaro.org/scheduler/job/18917#L3372

BTW, we had the 'sleep 10' in the beginning, but it was removed for
some reasons.

Change-Id: I9b40911568993dbcf3795c3d610ccd7cde7277d3
Signed-off-by: Chase Qi <chase.qi@linaro.org>