Update patch set 1
Patch Set 1:

> Senthil, thanks for your comment. The thing is, without the sleep
 > time, 'adb device' will print a serial number, but the following
 > adb command will fail. I suspect the 'adb device' get executed
 > before adbd restarting.

But, we have a 'adb wait-for-device' in between before calling 'adb devices', hence we will wait until a device becomes available. Also, I ve seen CTS tests ensuring device availability before proceeding to subsequent test cases.

 > We can add a loop to check if adb command returns correctly, but we
 > also need to set a timeout there, right?

Timeout to break the loop is good. But, the best solution is to not use any sleep or any loop.

 > From my point of view, I hope we can drop this work around, and run
 > 'adb root' in foreground, without '&' and without 'sleep 10'. This
 > works on xenial, with both latest and inhouse adb. Once CTT-489 get
 > solved, I will check this against stretch.

Please check for the problem existence at first place. Current LAVA code does not require any of these sorts of workaround since it exactly mimics a device connected directly to the host machine via udev rules. 

The removal of 'sleep 10' in the past was due to adding support for pyudev in LAVA which was capable of listening to udev add events. This was not due to any adb versions. As mentioned above, we have even better support now in LAVA.

Patch-set: 1