android: run 'adb root' in the foreground
Run 'adb root' in the background with a sleep time is a work around for
a specific adb version. It is not required in current LAVA code. Run
'adb root' directly works cross adb versions and distributions. It has
been tested on Ubuntu xenial, Debian jessie and stretch with inhouse and
latest adb.

jessie:
https://lkft.validation.linaro.org/scheduler/job/20107
https://lkft.validation.linaro.org/scheduler/job/20108

xenial:
https://lkft.validation.linaro.org/scheduler/job/20156
https://lkft.validation.linaro.org/scheduler/job/20155

Stretch:
https://lkft.validation.linaro.org/scheduler/job/20069
https://lkft.validation.linaro.org/scheduler/job/20140

Change-Id: Ibb3fa72b97729666656dbd749bc15b5446f37d5c
Signed-off-by: Chase Qi <chase.qi@linaro.org>