android: tradefed: handle timeout exception on tf shell exit
In some cases, TF shell refuse to exit with timeout in 60 seconds. Once
pexpect captured the pattern in index 0 or index 1, test should be
already finished. In the successful case, index 0, test report should be
already generated, refer to the following example. It should be safe to
kill the child process and continue to do the result parsing.

ResultReporter: Full Result: /lava-21588/0/tests/\
0_CtsFileSystemTestCases/automated/android/tradefed/android-cts/results\
/2017.08.28_21.28.08.zip

Change-Id: I5ffeea82d0904891db464dd08f89a2b199541e92
Signed-off-by: Chase Qi <chase.qi@linaro.org>