android: improve adb_debug_info()
Using 'lsusb' in LAVA-LXC directly is misleading. This is because from
within the container lsusb will show all usb devices that are connected
to host and there is no way to mask host devices from what is attached
to the container. Though the devices show up on lsusb, each device
should be added to the container via 'lxc-device add' command to make
it accessible from within the container(which is done by LAVA).

The current approach uses 'lsusb -D <device>' to get the information of
devices found in '/dev/bus/usb'

Change-Id: I28dc05f61685d995ac7fa9cc0e9831a936ef3a20
Signed-off-by: Chase Qi <chase.qi@linaro.org>