android: remove lsusb from adb_debug_info()
Quotes Senthil's comments:

"""
Do not use lsusb to see the usb devices connected within the container.
This is because from within the container lsusb will show all usb
devices that are connected to host and there is no way to mask host
devices from what is attached to the container. Though the devices show
up on lsusb, each device should be added to the container via
'lxc-device add' command to make it accessible from within the
container(which is done by LAVA).
"""

Change-Id: I28dc05f61685d995ac7fa9cc0e9831a936ef3a20
Signed-off-by: Chase Qi <chase.qi@linaro.org>