test-runner: add support for remote execution in LAVA
When '-l' or '--lava_run' option specified, test-runner will use
lava-test-case to send result to LAVA.

It might be possible to detect if it is running in LAVA at running time
by checking if 'lava-test-case' exists, but that is not reliable as
local environment can also has the util installed. So the above option
introduced with 'False' as default value.

Change-Id: Ief0c9e75d12c07832214d2d4fa30836df964615a
Signed-off-by: Chase Qi <chase.qi@linaro.org>