Update patch set 2
Patch Set 2:

The difference with this version and previous is that i removed the blogbench test because is not suitable for embedded devices due to a lot of usage of disk writes (disk space), is targeted for a server. The ddwrite test give us the same functionality to messure the performance of the embedded storage.

Patch-set: 2
Label: Code-Review=0