linux: sysbench: update required package names
On Debian Stretch, the lib package name changed to
default-libmysqlclient-dev. I checked on Ubuntu Artful Aardvark, it
ships both default-libmysqlclient-dev and libmysqlclient-dev. To make
sure both server and libraries come from the same codebase, changed
mysql-server to default-mysql-server as well.

This patch is not backwards compatible, let's keep it in erp branch
first. At the time we need it in master, we may be able to drop support
for old OS releases.

Change-Id: I2a2953e08591cd488900d5da0e1963e613362fe7
Signed-off-by: Chase Qi <chase.qi@linaro.org>