linux: sysbench: update required lib package name
On Debian Stretch, the lib package name changed to
default-libmysqlclient-dev. I checked on Ubuntu Artful Aardvark, it
ships both default-libmysqlclient-dev and libmysqlclient-dev.

This patch may be not backward compatible, lets keep it in erp branch
first. At the time we need it in master, we may be able to drop support
for old OS releases.

Change-Id: I2a2953e08591cd488900d5da0e1963e613362fe7
Signed-off-by: Chase Qi <chase.qi@linaro.org>