android tradefed: workaround for chown testuser home directory
This change is to workaround the problem happened here:
https://staging.validation.linaro.org/scheduler/job/194317#L5332

With adding more debug information, found that the problem
is caused by command of "chown testuser.testuser /home/testuser",
not sure why yet, but changing to create the home directory with
the useradd command could help to workaround that problem.

Also deleting that testuser before creating to make sure the useradd command
executed successfully, and delete that testuser after the test finished.

Change-Id: Icfd0dd1aa931a31d411e5e562c457d8c3ab007a7
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>