Update patch set 1
Patch Set 1:

If we want it in master, we should remove the test from other skipfiles too.

I wanted to cherry pick it to erp-18.01, but hit 'merge conflict', I guess that is  caused by the skipfile-rpk, as in the branch, the test not in the skipfile yet.

Patch-set: 1
Reviewer: Chase Qi <1000096@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=0