Update patch set 4
Patch Set 4:

i think i understand what you are trying to do, however I don't really like the format of the 'overlay' yaml file. the format ends up being quite different from the typical plan yaml format. it is a subset, but it's not very obvious to grasp.   (well to be honest, i don't really like the overlay test plan format, with requirements / name / tests , which doesn't seem very meaningful.. maybe we could use this excuse to define a new format for plans.. ;-)

do we really need to have 1 overlay that supports several 'boards' or could we instead have 1 overlay per board? If we had 1 overlay per board, we wouldn't have to detect the board name and we wouldn't have the 'weird' structure such as in the example, and the job or command line would simply list the test plan and any overlay to apply. Also that would generalize the concept of 'overlays', there is no specific reason why an overlay is 'board specific'. it could be distro specific, image specific, .. 

and the format could be:

tests:
   - path: automated/linux/disk-partitioning/disk-partitioning.yaml
      # Skip disk-partitioning test.
      skip: True
      parameters:

on the command line, we could provide a list of overlays to apply.

Patch-set: 4
Reviewer: Nicolas Dechesne <1000003@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=0