Update patch set 4
Patch Set 4:

Thanks for your comments, Nico! This is my initial attempt, it is open for ideas :)

> i think i understand what you are trying to do, however I don't
 > really like the format of the 'overlay' yaml file. the format ends
 > up being quite different from the typical plan yaml format. it is a
 > subset, but it's not very obvious to grasp.   (well to be honest, i
 > don't really like the overlay test plan format, with requirements /
 > name / tests , which doesn't seem very meaningful.. maybe we could
 > use this excuse to define a new format for plans.. ;-)
 > 

Test plan was designed in this way for reasons. @Milosz, can we strip off the 'name', 'owner' and 'category' and keep 'tests' section only to list all manual and automated tests? That will make post processing a bit easier.

The 'category' key probably should goes to test case yaml file. Test name is already covered in the file. I am not sure about the owner part.

 > do we really need to have 1 overlay that supports several 'boards'
 > or could we instead have 1 overlay per board? If we had 1 overlay
 > per board, we wouldn't have to detect the board name and we
 > wouldn't have the 'weird' structure such as in the example, and the
 > job or command line would simply list the test plan and any overlay
 > to apply. Also that would generalize the concept of 'overlays',
 > there is no specific reason why an overlay is 'board specific'. it
 > could be distro specific, image specific, ..
 > 
 > and the format could be:
 > 
 > tests:
 > - path: automated/linux/disk-partitioning/disk-partitioning.yaml
 > # Skip disk-partitioning test.
 > skip: True
 > parameters:
 > 
 > on the command line, we could provide a list of overlays to apply.

It was more ERP project considered, I wanted everything automated with one argument. Using more then one overlays when needed and leaving the choice to command line should be a better idea and easier to handle. I will give it a try.

Patch-set: 4