Update patch set 4
Patch Set 4:

> Test plan was designed in this way for reasons. @Milosz, can we strip off the 'name', 'owner' and 'category' and keep 'tests' section only to list all manual and automated tests? That will make post processing a bit easier.

> The 'category' key probably should goes to test case yaml file. Test name is already covered in the file. I am not sure about the owner part.

I'm not attached to the current format. There were reasons why it was done this way (tracking and approving requirements for RPB, categorizing results for reporting etc.) Most of these are no longer valid, so I see no reasons to keep the current format if it's not useful. There is a 'format' field in the test plan. We can come up with 'Linaro Test Plan v2' and redesign the whole thing.

Patch-set: 4
Label: Code-Review=0