Update patch set 4
Patch Set 4: Code-Review-1

> i think i understand what you are trying to do, however I don't
 > really like the format of the 'overlay' yaml file. the format ends
 > up being quite different from the typical plan yaml format. it is a
 > subset, but it's not very obvious to grasp.   (well to be honest, i
 > don't really like the overlay test plan format, with requirements /
 > name / tests , which doesn't seem very meaningful.. maybe we could
 > use this excuse to define a new format for plans.. ;-)
 > 
 > do we really need to have 1 overlay that supports several 'boards'
 > or could we instead have 1 overlay per board? If we had 1 overlay
 > per board, we wouldn't have to detect the board name and we
 > wouldn't have the 'weird' structure such as in the example, and the
 > job or command line would simply list the test plan and any overlay
 > to apply. Also that would generalize the concept of 'overlays',
 > there is no specific reason why an overlay is 'board specific'. it
 > could be distro specific, image specific, ..
 > 
 > and the format could be:
 > 
 > tests:
 > - path: automated/linux/disk-partitioning/disk-partitioning.yaml
 > # Skip disk-partitioning test.
 > skip: True
 > parameters:
 > 
 > on the command line, we could provide a list of overlays to apply.

Thanks Chase, this is a good idea. I agree with Nicolas that it should not be "board" specific. Rather, just have an argument to test-runner.py to specify the path to an amended test plan.

i.e. test-runner.py -a plans/d03.yaml
     test-runner.py -a plans/d05.yaml

And then the format of the amended test plan would not have a board name at top. It would just be a list of test modifications. You could have some name such as "amendments:", so that in the future other things could be specified. For example, d05.yaml could contain:

amendments:
    - path: automated/linux/disk-partitioning/disk-partitioning.yaml
      # Skip disk-partitioning test.
      skip: True
      parameters:
    - path: automated/linux/network-basic/network-basic.yaml
      skip: False
      # Test the second network interface.
      parameters:
          INTERFACE: enahisic2i1

Determining which test amendment file belongs to which environment is the job of whatever is calling test-runner (jenkins, lava, ansible, etc).

Patch-set: 4
Label: Code-Review=-1