Update patch set 2
Patch Set 2: Code-Review+2

This looks good to me and I think it's important to get it in quickly. That said, I think it would be better for next time to split this into one commit that is doing all the file reorganization and another for new skips.

Patch-set: 2
Label: Code-Review=+2