kselftest: skipfile: skip run_vmtests on 4.4 and 4.9
The source code of this test case has been changed largely.
We have not found any graceful way to make them work all together
to be compatible with old and new kernel.

So we'd like to abandon it.

Ref:
LKFT: linux-stable-rc-4.9 and 4.4: kselftest: run_vmtests:
userfaultfd, userfaultfd_hugetlb and userfaultfd_shmem test cases failed
https://bugs.linaro.org/show_bug.cgi?id=3171

Change-Id: Ibf9d870251a05782f0d5e5c6454ca447fc4bf471
Signed-off-by: Naresh Kamboju <naresh.kamboju@linaro.org>