Update patch set 2
Patch Set 2:

I tested skipgen a bit and it doesn't seem to work as expected. I might not understand how it's supposed to work :)
When I use 'all' I expect all possible envs/boards/branches to be covered. Instead I'm only getting the list of skipped tests that apply to 'all'. So if test applies to 'production' and 'staging' it doesn't automatically apply to 'all'. Is that expected? It doesn't look intuitive to me.

Patch-set: 2
Label: Code-Review=0