LTP: Convert LKFT skipfiles to a yaml skipfile
skipfile-lkft.yaml encompasses all of the lkft skipfiles in this
directory as well as the skipfiles/ directory. They may be removed once
build jobs are updated to use a generated skipfile.

To generate a skipfile from skipfile-lkft.yaml, use skipgen, which is
included in this repository at /bin/<arch>/skipgen. ltp.sh will be
expanded to support skipgen separately.

The actual contents have been derived from the various skipfiles. In
cases of inconsistency, the skip was expanded. For example,
netns_breakns_ip_ipv4_ioctl is currently skipped on 4.4 hikey, 4.4 juno,
and all branches of x15 and x86. In skipfile-lkft.yaml, it's skipped for
all branches on all boards.

No other modifications to the skipfiles have been implemented. There are
obvious skips that need adjusting, but that will be done subsequently.
This review is just a one to one translation of the skipfiles into one
yaml skipfile. If you notice things that are wrong or shouldn't be
listed anymore - good! That's the point of this effort 😀

Lastly, teach validate.py about skipgen yaml files. Validate them by
running skipgen itself against them and verifying that a skiplist of at
least 1 skip is produced.

Change-Id: Ic46af16d36dfd30ed0876808f916318cddd44fa2
Signed-off-by: Dan Rue <dan.rue@linaro.org>