[RFC] LTP: Add skipgen support to ltp.sh
Create a helper function in sh-test-lib to handle the mechanics of
generating a skipfile, that will be re-usable by kselftest.

Add the following arguments to ltp.sh:
    [-Y skipfile-yaml]
      [-b board]
      [-g branch]
      [-e environment]

Where skipfile-yaml is a yaml formatted skipfile. Board, branch, and
environment are optional arguments that get passed to skipgen to
generate a skipfile at runtime.

The resulting skipfile is automatically printed to stdout after being
generated.

Change-Id: Ib45d11acfa85241dbe5dfa7173e75cb7754a6892
Signed-off-by: Dan Rue <dan.rue@linaro.org>