android/tradefed: don't exit when test run/module fails to complete
The second pattern is used to match the end of a failed test run that
didn't generate test result, but it matches a failed test module too. In
the second case, the whole run will be force killed, which isn't
desired.

Ideally, we should compare a failed run and a failed module to find a
better pattern/patterns to distinguish between them. However, we cannot
duplicate the failed run mentioned above anymore.

The current workaround is mark the whole test run as fail when a run or
a module failed to complete and let test continue. Once we hit the above
failed run again, we may be able to find a better solution.

Change-Id: I506e0ead4d1a937423fe797f51b7decf1eb343a2
Signed-off-by: Chase Qi <chase.qi@linaro.org>