[RFC] LTP: Add skipgen skipfile support to ltp
Create a helper function in sh-test-lib to handle the mechanics of
generating a skipfile, that will be re-usable by kselftest.

Add the following arguments to ltp.sh:
    [-b board]
    [-g branch]
    [-e environment]

When the skipfile argument passed via -S ends in ".yaml", a skipfile
will be generated using skipgen. Board, branch, and environment are
optional arguments that get passed to skipgen.

The resulting skipfile is automatically printed to stdout after being
generated.

Change-Id: Ib45d11acfa85241dbe5dfa7173e75cb7754a6892
Signed-off-by: Dan Rue <dan.rue@linaro.org>