Update patch set 3
Patch Set 3:

> It looks OK to me. How do we roll it out to avoid breakages in
 > production?

It's backward compatible so it should continue to work as jobs move over.

There's one more thing I want to implement before merging. Right now, if a url is given, it is assumed to be a regular skipfile, but it should be able to handle a yaml skipfile or a regular skipfile. Once I do that I'll re-test all the code paths and then we can merge.

Patch-set: 3
Label: Code-Review=0