plan: add option to render single file plans
Added -s (--single-file-output) option to testplan2html.py script. This
allows to render test steps and expected results in the same file as the
other bits of test plan. The generic test metadata is omited in this
case. Parameters are combination of predefined test parameters and
parameters defined in the test plan. Parameters from the test plan take
precedence.

Change-Id: I99291aaa2744603d88bda337c3f2e2e79fedb89a
Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>