Update patch set 2
Patch Set 2: Code-Review-1

This is wrong approach. If the total number of failures is the culprit, the patch has to be re-designed. I agree with Chase that showing partial failures makes it confusing. Maybe we should add a parameter that limits the number of displayed failures? Anyway, the code is generic and the argument that it works in special case is inappropriate.

Patch-set: 2
Label: Code-Review=-1