Update patch set 2
Patch Set 2:

> This is wrong approach. If the total number of failures is the
 > culprit, the patch has to be re-designed. I agree with Chase that
 > showing partial failures makes it confusing. Maybe we should add a
 > parameter that limits the number of displayed failures? Anyway, the
 > code is generic and the argument that it works in special case is
 > inappropriate.

The output of failed test cases is just for initial debugging/analysis, it does not affect the result reported.

To do deep investigation, uploading the result to archive site should be the right solution, but that needs open the job to find the archive url, and download them, and check the files, it's a little complex for an initial check. With this output information, we just need to open the log, and would get the failed test cases and its short debug message.

Here limited with hard coded number 200 is enough for debug, changing the interface for debugging code is not necessary I think. but if you think it will be better to add a parameter to pass the number, we could change the interface to have it.

Patch-set: 2