Update patch set 2
Patch Set 2:

I claim this is one-off special case because you keep on insisting it works for you in this and that test job. If you want a general purpose solution, please introduce a parameter (that's off by default) and make sure the test job always passes by limiting the number of failed tests per job to some number (preferably a parameter as well).

Patch-set: 2