LKFT: kselftest: Introduce yaml skipfile
This skipfile is consumable using skipgen
(https://github.com/Linaro/skipgen), included in this repo in
/automated/bin/<arch>/skipgen.

The skips in this file accurately reflect the skips in skipfiles/, with
the following considerations. When there were inconsistencies, new skips
were added to resolve, or multiple skip blocks were used. In no case
were skips removed from a board/branch combination.

The inconsistencies, missing data, etc in this file are due to
inconsistencies in skipfiles/, and will be resolved in a second pass
once migration to the yaml skipfile is complete and all the skipfiles/
are removed. That said, this file was generated manually (and validated
computationally), so there may also be errors introduced.

Verified with the following two scripts:

    # Compare complete lists of skips to find any that are entirely missing. Note
    # that this catches the 'rpk' skips, which should be ignored.
    skipgen skipfile-lkft.yaml | sort > some
    cat skipfiles/* | grep -v ^$ | grep -v ^# | awk '{print $1}' | sort | uniq > all
    diff some all

    # Check each board/branch combination for matching skiplists
    for board in dragonboard-410c hi6220-hikey juno-r2 x15 x86 qemu_arm64 qemu_x86_64; do
        for branch in 4.15 4.14 4.9 4.4; do
            echo $board $branch
            skipgen --board $board --branch $branch --environment production skipfile-lkft.yaml | sort > 1
            cat skipfiles/skiplist_${board}_${branch} | grep -v ^$ | grep -v ^# | awk '{print $1}' | sort | uniq > 2
            if ! diff 1 2; then
                read
            fi
        done
    done

Change-Id: Id06bf6bfa16c9f103bed254bbbf6a0242e91c65b
Signed-off-by: Dan Rue <dan.rue@linaro.org>